begin filter page helper
